Always use random filenames for images
Gwicke had a concern when he did a security look through that filenames
could be used to cause damage to the latex engine / the box. Previously
we escaped some stuff and then fell back to using the tmp module -
this just uses the tmp module directly avoiding all escaping and collision
problems.

Change-Id: I3520ae6ab7f1d68b6d285c023ef08e547193a34f